requirements.txt
django
pillow
djangocms-admin-style
psycopg2
django-ckeditor
django-allauth